==> portfolio-ev-fy20/02 Data file generation.Rmd <==
---
title: "02 Data file generation"
output:
  html_notebook:
    toc: TRUE
    toc_float:
      collapsed: FALSE
---

# Goal

Explore using featuretoolsR to create a large data frame for use in modeling.

```{r setup, message = FALSE, warning = FALSE}
library(tidyverse)
library(lubridate)
library(wranglR)
library(readxl)
library(foreach)
library(reticulate)
library(nanotime)
library(featuretoolsR)

source(file = 'code/featuretools helper functions.R')
```

# Data import

```{r}
fileloader <- function(filename, sheetnum) {
  read_xlsx(
    path = paste0('data/', filename)
    , sheet = sheetnum
    , guess_max = 1E7
  )
}

# Base table
households <- fileloader('2020-03-16 Engagement data.xlsx', 1)
# Biodata tables
bio_phone <- fileloader('2020-03-16 Addr phone.xlsx', 1)
bio_address <- fileloader('2020-03-16 Addr phone.xlsx', 2)
bio_employment <- fileloader('2020-03-16 employment email.xlsx', 1)
bio_email <- fileloader('2020-03-16 employment email.xlsx', 2)
# Engagement tables
eng_committee <- fileloader('2020-03-16 Engagement data.xlsx', 2)
eng_event <- fileloader('2020-03-16 Engagement data.xlsx', 3)
eng_activity <- fileloader('2020-03-16 Engagement data.xlsx', 4)
# Giving tables
giv_transactions <- fileloader('2020-03-16 Gift transactions.xlsx', 1)
giv_payments <- fileloader('2020-03-16 Gift data.xlsx', 1)
giv_giftclub <- fileloader('2020-03-16 Gift data.xlsx', 2)
# Prospect tables
prs_evaluation <- fileloader('2020-03-16 Prospect data.xlsx', 1)
prs_contact <- fileloader('2020-03-16 Prospect data.xlsx', 2)
prs_program <- fileloader('2020-03-16 Prospect data.xlsx', 3)
prs_assignment <- fileloader('2020-03-16 Prospect data.xlsx', 4)
```

# Data cleanup

```{r}
households <- households %>%
  # Factors
  mutate(
    RECORD_STATUS_CODE = factor(RECORD_STATUS_CODE)
    , HOUSEHOLD_PROGRAM = factor(HOUSEHOLD_PROGRAM)
    , HOUSEHOLD_PROGRAM_GROUP = factor(HOUSEHOLD_PROGRAM_GROUP)
  ) %>%
  # Numeric
  mutate_at(
    vars(ends_with('_YEAR'))
    , list(as.numeric)
  )

summary(households)
```

```{r}
bio_address <- bio_address %>%
  mutate(
    ADDR_TYPE_CODE = factor(ADDR_TYPE_CODE)
    , ADDR_TYPE = factor(ADDR_TYPE)
    , BUSINESS_IND = BUSINESS_IND == 'Y'
    , HBS_CODE = factor(HBS_CODE)
    , ADDR_STATUS_CODE = factor(ADDR_STATUS_CODE)
  )

summary(bio_address)
```

Note that any start year past 2020 is suspect.

```{r}
bio_email <- bio_email %>%
  mutate(
    EMAIL_TYPE_CODE = factor(EMAIL_TYPE_CODE)
    , EMAIL_TYPE = factor(EMAIL_TYPE)
    , EMAIL_STATUS_CODE = factor(EMAIL_STATUS_CODE)
  )

summary(bio_email)
```

```{r}
bio_employment <- bio_employment %>%
  mutate(
    JOB_STATUS_CODE = factor(JOB_STATUS_CODE)
    , SELF_EMPLOY_IND = SELF_EMPLOY_IND == 'Y'
    , MATCHING_STATUS_IND = MATCHING_STATUS_IND == 'Y'
  )

summary(bio_employment)
```

```{r}
bio_phone <- bio_phone %>%
  mutate(
    TELEPHONE_TYPE_CODE = factor(TELEPHONE_TYPE_CODE)
    , TELEPHONE_TYPE = factor(TELEPHONE_TYPE)
    , BUSINESS_IND = BUSINESS_IND == 'Y'
    , CURRENT_IND = CURRENT_IND == 'Y'
    , TELEPHONE_STATUS_CODE = factor(TELEPHONE_STATUS_CODE)
  )

summary(bio_phone)
```

Note to self - re-pull telephone data with additional types (done). Those 1902 dates do look pretty fishy.

```{r}
eng_activity <- eng_activity %>%
  mutate(
    ACTIVITY_DESC = factor(ACTIVITY_DESC)
    , KSM_ACTIVITY = !is.na(KSM_ACTIVITY)
    , ACTIVITY_PARTICIPATION_CODE = factor(ACTIVITY_PARTICIPATION_CODE)
  )

summary(eng_activity)
```

Clean up start/stop FY.

```{r}
eng_committee <- eng_committee %>%
  mutate(
    COMMITTEE_CODE = factor(COMMITTEE_CODE)
    , COMMITTEE_DESC = factor(COMMITTEE_DESC)
    , COMMITTEE_ROLE_CODE = factor(COMMITTEE_ROLE_CODE)
    , COMMITTEE_ROLE = factor(COMMITTEE_ROLE)
    , KSM_COMMITTEE = !is.na(KSM_COMMITTEE)
    , COMMITTEE_STATUS = factor(COMMITTEE_STATUS)
  )

summary(eng_committee)
```

```{r}
eng_event <- eng_event %>%
  mutate(
    EVENT_NAME = factor(EVENT_NAME)
    , KSM_EVENT = !is.na(KSM_EVENT)
    , EVENT_TYPE = factor(EVENT_TYPE)
  )

summary(eng_event)
```

```{r}
giv_giftclub <- giv_giftclub %>%
  mutate(
    GIFT_CLUB = factor(GIFT_CLUB)
    , GIFT_CLUB_CODE = factor(GIFT_CLUB_CODE)
    , GC_CATEGORY = factor(GC_CATEGORY)
  )

summary(giv_giftclub)
```

```{r}
giv_payments <- giv_payments %>%
  mutate(
    PLEDGE_FY = as.numeric(PLEDGE_FY)
  )

summary(giv_payments)
```

```{r}
giv_transactions <- giv_transactions %>%
  mutate(
    TRANSACTION_TYPE = factor(TRANSACTION_TYPE)
    , TX_GYPM_IND = factor(TX_GYPM_IND)
    , ASSOCIATED_DESC = factor(ASSOCIATED_DESC)
    , PAYMENT_TYPE = factor(PAYMENT_TYPE)
    , ALLOC_SHORT_NAME = factor(ALLOC_SHORT_NAME)
    , AF_FLAG = !is.na(AF_FLAG)
    , CRU_FLAG = !is.na(CRU_FLAG)
    , PROPOSAL_ID = factor(PROPOSAL_ID)
  )

summary(giv_transactions)
```

```{r}
prs_assignment <- prs_assignment %>%
  mutate(
    ASSIGNMENT_TYPE = factor(ASSIGNMENT_TYPE)
    , ASSIGNMENT_TYPE_DESC = factor(ASSIGNMENT_TYPE_DESC)
    , ASSIGNMENT_REPORT_NAME = factor(ASSIGNMENT_REPORT_NAME)
    , COMMITTEE_DESC = factor(COMMITTEE_DESC)
  )

prs_assignment %>% select(-ASSIGNMENT_REPORT_NAME) %>% summary()
```

No meaningful COMMITTEE_DESC data.

```{r}
prs_contact <- prs_contact %>%
  mutate(
    CONTACT_CREDIT_TYPE = factor(CONTACT_CREDIT_TYPE)
    , CONTACT_CREDIT_DESC = factor(CONTACT_CREDIT_DESC)
    , ARD_STAFF = !is.na(ARD_STAFF)
    , FRONTLINE_KSM_STAFF = !is.na(FRONTLINE_KSM_STAFF)
    , CONTACT_TYPE_CATEGORY = factor(CONTACT_TYPE_CATEGORY)
    , VISIT_TYPE = factor(VISIT_TYPE)
  )

summary(prs_contact)
```

```{r}
prs_evaluation <- prs_evaluation %>%
  mutate(
    EVALUATION_TYPE = factor(EVALUATION_TYPE)
    , EVALUATION_DESC = factor(EVALUATION_DESC)
    , RATING_CODE = factor(RATING_CODE)
    , RATING_DESC = factor(RATING_DESC)
  )

summary(prs_evaluation)
```

```{r}
prs_program <- prs_program %>%
  mutate(
    PROGRAM_CODE = factor(PROGRAM_CODE)
    , PROGRAM = factor(PROGRAM)
  )

summary(prs_program)
```

# Quick test

Test features with biodata only.

```{r}
hh <- households %>%
  select(
    -HOUSEHOLD_RPT_NAME
    , -HOUSEHOLD_SPOUSE_ID
    , -HOUSEHOLD_SPOUSE_RPT_NAME
    , -RECORD_STATUS_CODE
    , -DATA_AS_OF
  )
```

Create entities.

```{r}
es <- as_entityset(
  hh # data frame
  , entity_id = 'household' # name of entity
  , index = 'HOUSEHOLD_ID' # unique identifier
  , id = 'household' # name of entity set
) %>%
  # Bio entities
  add_entity(
    bio_address %>% select(-START_DT, -STOP_DT, -DATE_MODIFIED, -ADDR_TYPE_CODE)
    , entity_id = 'bio_address'
    , make_index = TRUE
    , index = 'bio_address_idx'
    , time_index = 'DATE_ADDED' # when the record can first be used
  ) %>%
  add_entity(
  bio_email %>% select(-EMAIL_TYPE_CODE, -START_DT, -STOP_DT)
  , entity_id = 'bio_email'
  , make_index = TRUE
  , index = 'bio_email_idx'
  , time_index = 'DATE_ADDED' # when the record can first be used
  ) %>%
  add_entity(
  bio_employment %>% select(-START_DT, -STOP_DT, -JOB_TITLE, -EMPLOYER_NAME)
  , entity_id = 'bio_employment'
  , make_index = TRUE
  , index = 'bio_employment_idx'
  , time_index = 'DATE_ADDED' # when the record can first be used
  ) %>%
  add_entity(
  bio_phone %>% select(-TELEPHONE_TYPE_CODE, -START_DT, -STOP_DT)
  , entity_id = 'bio_phone'
  , make_index = TRUE
  , index = 'bio_phone_idx'
  , time_index = 'DATE_ADDED' # when the record can first be used
)
```

Create relationships.

```{r}
es <- es %>%
  add_relationship(
    parent_set = 'household'
    , child_set = 'bio_address'
    , parent_idx = 'HOUSEHOLD_ID'
    , child_idx = 'HOUSEHOLD_ID'
  ) %>%
  add_relationship(
    parent_set = 'household'
    , child_set = 'bio_email'
    , parent_idx = 'HOUSEHOLD_ID'
    , child_idx = 'HOUSEHOLD_ID'
  ) %>%
  add_relationship(
    parent_set = 'household'
    , child_set = 'bio_employment'
    , parent_idx = 'HOUSEHOLD_ID'
    , child_idx = 'HOUSEHOLD_ID'
  ) %>%
  add_relationship(
    parent_set = 'household'
    , child_set = 'bio_phone'
    , parent_idx = 'HOUSEHOLD_ID'
    , child_idx = 'HOUSEHOLD_ID'
  )
```

```{r}
print(es)
```

```{r}
# Use base R data
dfs_r <- es %>% dfs(
  target_entity = 'household'
  , agg_primitives = c('count')
  , trans_primitives = c('month', 'year')
  , max_depth = 2
  , verbose = TRUE
)
```

```{r, rows.print = 20}
print_features(dfs_r)
```

We see that dfs created counts for the total number of address, email, etc. entries on the table per household_id. However, it would be more useful to get counts broken down by other criteria -- for example, how many historical home addresses, active emails, etc.

# "Interesting values" function

The featuretools "interesting values" functionality creates new aggregation fields for each of a predefined set of factor levels -- or just the n most common.

```{python}
import featuretools as ft

# Import R data
py_es = r.es

# Add interesting values tabulation; here I'm looking at up to the 6 most common levels for each table.
py_es.add_interesting_values(max_values = 6, verbose = True)
```

```{r}
# Import data from python
dfs_py <- py$py_es %>% dfs(
  target_entity = 'household'
  , agg_primitives = c('count')
  , trans_primitives = c('month', 'year')
  , max_depth = 2
  , verbose = TRUE
)
```

```{r, rows.print = 20}
print_features(dfs_py)
```

Here we see the interesting values functionality added counts specifically by contact status, type, etc. This took only a few seconds, which definitely beats doing this for dozens of fields by hand.

# Full data test

Clean up the environment.

```{r}
rm(list = ls())
source(file = 'code/featuretools helper functions.R')
```

## Load data

```{r, warning = FALSE}
# Change this to the date the data files were generated
data_dt <- '2020-04-08'
source(file = 'code/data xlsx import.R')
```

## Check data

```{r}
source(file = 'code/data validation.R')
```

Fix the start date issues.

```{r}
catracks$giv_giftclub %>% select(START_DT) %>% summary()
catracks$prs_evaluation %>% select(EVAL_DT) %>% summary()
catracks$prs_contact %>% select(CONTACT_DATE) %>% summary()
```

Missing gift club dates can be filled in with start year or date added. For evaluation and contact, it's difficult to think of an equally principled way to automatically impute data (start date is the contact date approximately 0% of the time). Since there are only a handful of rows I'm leaning toward simply deleting them.

```{r}
# Fill in start dates with either the valid date part (e.g. year only), or the date added
catracks$giv_giftclub <- catracks$giv_giftclub %>%
  mutate(
    START_DT_CALC = case_when(
      !is.na(START_DT) ~ as.Date(START_DT)
      , substr(GIFT_CLUB_START_DATE, 1, 4) == '0000' ~ as.Date(DATE_ADDED)
      , TRUE ~ ToDateCleaner(GIFT_CLUB_START_DATE)
    )
  )


# Set new attributes
attr(catracks$giv_giftclub, 'time_index') <- 'START_DT_CALC'

# Remove eval and contact without a valid date
catracks$prs_evaluation <- catracks$prs_evaluation %>%
  filter(!is.na(EVAL_DT))
catracks$prs_contact <- catracks$prs_contact %>%
  filter(!is.na(CONTACT_DATE))
```

```{r}
source(file = 'code/data validation.R')
```

## Build entity set

Modeling data includes a subset of fields from the catracks dataset.

```{r}
mdat <- catracks

mdat$households <- mdat$households %>%
  select(-HOUSEHOLD_SPOUSE_ID, -HOUSEHOLD_SPOUSE_RPT_NAME, -DATA_AS_OF)
# Biodata
mdat$bio_address <- mdat$bio_address %>%
  select(-START_DT, -STOP_DT, -DATE_MODIFIED, -ADDR_TYPE_CODE)
mdat$bio_email <- mdat$bio_email %>%
  select(-EMAIL_TYPE_CODE, -START_DT, -STOP_DT)
mdat$bio_employment <- mdat$bio_employment %>%
  select(-START_DT, -STOP_DT, -JOB_TITLE, -EMPLOYER_NAME)
mdat$bio_phone <- mdat$bio_phone %>%
  select(-TELEPHONE_TYPE_CODE)
# Engagement
mdat$eng_activity <- mdat$eng_activity %>%
  select(-ID_NUMBER, -START_DT, -STOP_DT)
mdat$eng_committee <- mdat$eng_committee %>%
  select(-ID_NUMBER, -COMMITTEE_CODE, -COMMITTEE_ROLE_CODE, -COMMITTEE_ROLE_XSEQUENCE)
mdat$eng_event <- mdat$eng_event %>%
  select(-ID_NUMBER, -EVENT_ID)
# Giving
mdat$giv_giftclub <- mdat$giv_giftclub %>%
  select(-GIFT_CLUB_CODE, -GIFT_CLUB_START_DATE, -GIFT_CLUB_END_DATE, -STOP_DT)
mdat$giv_transactions <- mdat$giv_transactions %>%
  select(-TX_NUMBER, -TX_SEQUENCE, -MATCHED_TX_NUMBER, -PLEDGE_NUMBER)
# Prospect
mdat$prs_assignment <- mdat$prs_assignment %>%
  select(-ASSIGNMENT_TYPE, COMMITTEE_DESC)
mdat$prs_contact <- mdat$prs_contact %>%
  select(-CONTACT_CREDIT_TYPE)
mdat$prs_evaluation <- mdat$prs_evaluation %>%
  select(-EVALUATION_TYPE, -RATING_CODE)
mdat$prs_program <- mdat$prs_program %>%
  select(-PROGRAM_CODE, -START_DATE, -STOP_DATE)
```


```{r}
# Function to fill in entity information
make_entityset <- function(df, df_name, debug = FALSE) {
  if (debug) {print(paste('Adding', df_name))}
  r_entityset %>%
    add_entity(
      df
      , entity_id = df_name
      , make_index = TRUE
      , index = paste0(df_name, '_idx')
      , time_index = attr(df, 'time_index')
    )
}

# Create entity set
r_entityset <- as_entityset(
  mdat$households # data frame
  , entity_id = 'households' # name of entity
  , index = 'HOUSEHOLD_ID' # unique identifier
  , time_index = attr(mdat$households, 'time_index') # when the record can first be used
  , id = 'households' # name of entity set
)

# Load all entities
r_entityset <- mapply(
  make_entityset
  , df = within(mdat, rm('households'))
  , df_name = names(within(mdat, rm('households')))
)[[1]]
```

Create relationships.

```{r}
# Function to fill in relationships
make_relationships <- function(df_name, debug = FALSE) {
  if (debug) {print(df_name)}
  r_entityset %>%
    add_relationship(
      parent_set = 'households'
      , child_set = df_name
      , parent_idx = 'HOUSEHOLD_ID'
      , child_idx = 'HOUSEHOLD_ID'
    )
}

r_entityset <- sapply(
  within(mdat, rm('households')) %>% names()
  , FUN = make_relationships
)[[1]]
```

```{r}
print(r_entityset)
```

## Giving data test

Quick test with a giving dataset.

```{r}
# Create entity set and relationships
giving_entityset <- as_entityset(
  mdat$households %>% tail(10E3) # data frame
  , entity_id = 'households' # name of entity
  , index = 'HOUSEHOLD_ID' # unique identifier
  , time_index = attr(mdat$households, 'time_index') # when the record can first be used
  , id = 'households' # name of entity set
) %>% add_entity(
  mdat$giv_transactions %>% filter(HOUSEHOLD_ID %in% {mdat$households %>% tail(10E3) %>% select(HOUSEHOLD_ID) %>% unlist()})
  , entity_id = 'giv_transactions'
  , make_index = TRUE
  , index = 'giv_transactions_idx'
  , time_index = attr(mdat$giv_transactions, 'time_index')
) %>% add_relationship(
  parent_set = 'households'
  , child_set = 'giv_transactions'
  , parent_idx = 'HOUSEHOLD_ID'
  , child_idx = 'HOUSEHOLD_ID'
)
# Check results
print(giving_entityset)
```

```{python}
import featuretools as ft
import pandas as pd
import numpy as np

# Import R data
py_giv = r.giving_entityset

# Add interesting values tabulation; here I'm looking at up to the 6 most common levels for each table.
py_giv.add_interesting_values(max_values = 6, verbose = False)
```

```{r}
# Depth 2, lots of combinations
dfs_giv <- py$py_giv %>% dfs(
   target_entity = 'households'
   , agg_primitives = c('count', 'sum', 'std', 'mean', 'max', 'min', 'median', 'first', 'last', 'percent_true')
   , trans_primitives = c('cum_sum', 'cum_max', 'month', 'year', 'subtract_numeric', 'divide_numeric', 'time_since_previous')
   , max_depth = 2
   , verbose = TRUE
)
```

```{r, rows.print = 40}
# Results -- commented out due to size
# print_features(dfs_giv)
# dfs_giv[[1]] %>% filter(`COUNT(giv_transactions)` > 0) %>% View()
```

## Large interesting values test

First, add interesting values in Python.

```{python}
# Import R data
py_es = r.r_entityset

# Add interesting values tabulation; here I'm looking at up to the 6 most common levels for each table.
py_es.add_interesting_values(max_values = 6, verbose = False)
```

```{r}
# Import the data back into R
py_entityset <- py$py_es
```

```{r}
# Quick test using the full data from Python and cutoff_time
# Should be about a minute on a strong multicore system
dfs_py <- py_entityset %>% dfs(
  target_entity = 'households'
  , agg_primitives = c('count')
  , trans_primitives = c('year')
  , max_depth = 1
  , verbose = TRUE
)
```

```{r, rows.print = 40}
# Results -- commented out due to size
# print_features(dfs_py)
# dfs_py[[1]] %>% View()
```

## Time filter test

Couldn't get this to work in R (datetime format issues) so I created helper functions instead to pre-filter data before creating an entityset.
