==> portfolio-ev-fy20/03 Data file output.Rmd <==
---
title: "03 Data file output"
output:
  html_notebook:
    toc: TRUE
    toc_float:
      collapsed: FALSE
---

# Goal

Create a set of modeling-ready, time-sliced data files with featuretoolsR.

# Setup

```{r setup, message = FALSE, warning = FALSE}
library(tidyverse)
library(lubridate)
library(wranglR)
library(readxl)
library(foreach)
library(reticulate)
library(nanotime)
library(featuretoolsR)

source(file = 'code/featuretools helper functions.R')

# Parameters
n <- 4
```

# Load and check data

```{r, warning = FALSE}
data_dt <- '2020-04-08'
source(file = 'code/data xlsx import.R')
source(file = 'code/data cleanup.R')
source(file = 'code/data validation.R')
```

# Set up Python session

```{python}
import featuretools as ft
```

# Create chunked entity sets

```{r, message = FALSE}
# Chunk the data into n (approximately) equal-sized bins
catracks_chunked <- chunk_datalist(
  catracks
  , master_table_name = 'households'
  , master_idx_name = 'HOUSEHOLD_ID'
  , chunks = n
  , seed = 123
)
```

```{r}
cutoff_dt <- ymd('20180831')
n_interesting_values <- 6L

entities_fy18 <- foreach(
  i = 1:length(catracks_chunked)
  , .combine = list
  , .multicombine = TRUE) %do% {
    # Create entityset
    a <- catracks_chunked[[i]] %>%
      entityset_create(
        entityset_name = paste0('chunk', i)
        , cutoff_dt = cutoff_dt
        , master_entity = 'households'
        , master_idx = 'HOUSEHOLD_ID'
        , debug = TRUE
      )
    # Insert interesting values
    a$add_interesting_values(max_values = n_interesting_values)
    # Return result as a list item
    return(a)
  }
```

```{r}
foreach (i = 1:n) %do% {
  print(entities_fy18[[i]])
}
```

Comparing this to the full dataset:

```{r, message = FALSE}
catracks %>%
  entityset_create(
    entityset_name = paste0('chunk', i)
    , cutoff_dt = cutoff_dt
    , master_entity = 'households'
    , master_idx = 'HOUSEHOLD_ID'
    , debug = FALSE
  ) %>%
  print()
```

Each chunk does contain about `r (1/n) %>% scales::percent() %>% I()` of the total data, as expected.

# Full run

```{r, cache = TRUE}
# Full run (slow)
dfs_output_fy18_1 <- entities_fy18[[1]] %>%
  dfs(
   target_entity = 'households'
   , agg_primitives = c('count', 'sum', 'std', 'mean', 'max', 'min', 'median', 'first', 'last', 'percent_true')
   , trans_primitives = c('cum_sum', 'cum_max', 'month', 'year', 'subtract_numeric', 'divide_numeric', 'time_since_previous')
   , max_depth = 2
   , verbose = TRUE
)

save('dfs_output_fy18_1', file = 'data/output/dfs_output_fy18_1.Rdata')
```

```{r, cache = TRUE}
# Full run (slow)
dfs_output_fy18_2 <- entities_fy18[[2]] %>%
  dfs(
   target_entity = 'households'
   , agg_primitives = c('count', 'sum', 'std', 'mean', 'max', 'min', 'median', 'first', 'last', 'percent_true')
   , trans_primitives = c('cum_sum', 'cum_max', 'month', 'year', 'subtract_numeric', 'divide_numeric', 'time_since_previous')
   , max_depth = 2
   , verbose = TRUE
)

save('dfs_output_fy18_2', file = 'data/output/dfs_output_fy18_2.Rdata')
```

```{r, cache = TRUE}
# Full run (slow)
dfs_output_fy18_3 <- entities_fy18[[3]] %>%
  dfs(
   target_entity = 'households'
   , agg_primitives = c('count', 'sum', 'std', 'mean', 'max', 'min', 'median', 'first', 'last', 'percent_true')
   , trans_primitives = c('cum_sum', 'cum_max', 'month', 'year', 'subtract_numeric', 'divide_numeric', 'time_since_previous')
   , max_depth = 2
   , verbose = TRUE
)

save('dfs_output_fy18_3', file = 'data/output/dfs_output_fy18_3.Rdata')
```

```{r, cache = TRUE}
# Full run (slow)
dfs_output_fy18_4 <- entities_fy18[[4]] %>%
  dfs(
   target_entity = 'households'
   , agg_primitives = c('count', 'sum', 'std', 'mean', 'max', 'min', 'median', 'first', 'last', 'percent_true')
   , trans_primitives = c('cum_sum', 'cum_max', 'month', 'year', 'subtract_numeric', 'divide_numeric', 'time_since_previous')
   , max_depth = 2
   , verbose = TRUE
)

save('dfs_output_fy18_4', file = 'data/output/dfs_output_fy18_4.Rdata')
```

# Combine output files into one dataframe

```{r}
# Use pagefile
utils::memory.limit(128000)

# Load data
load('data/output/dfs_output_fy18_1.Rdata')
load('data/output/dfs_output_fy18_2.Rdata')
load('data/output/dfs_output_fy18_3.Rdata')
load('data/output/dfs_output_fy18_4.Rdata')

# Keep A, C, L record statuses
dfs_output_fy18_1[[1]] <- dfs_output_fy18_1[[1]] %>% filter(RECORD_STATUS_CODE %in% c('A', 'C', 'L'))
dfs_output_fy18_2[[1]] <- dfs_output_fy18_2[[1]] %>% filter(RECORD_STATUS_CODE %in% c('A', 'C', 'L'))
dfs_output_fy18_3[[1]] <- dfs_output_fy18_3[[1]] %>% filter(RECORD_STATUS_CODE %in% c('A', 'C', 'L'))
dfs_output_fy18_4[[1]] <- dfs_output_fy18_4[[1]] %>% filter(RECORD_STATUS_CODE %in% c('A', 'C', 'L'))

# Append HHID
dfs_output_fy18_1[[1]]$HOUSEHOLD_ID <- rownames(dfs_output_fy18_1[[1]])
dfs_output_fy18_2[[1]]$HOUSEHOLD_ID <- rownames(dfs_output_fy18_2[[1]])
dfs_output_fy18_3[[1]]$HOUSEHOLD_ID <- rownames(dfs_output_fy18_3[[1]])
dfs_output_fy18_4[[1]]$HOUSEHOLD_ID <- rownames(dfs_output_fy18_4[[1]])

gc()
```

```{r}
dfs_output_fy18 <- bind_rows(
    dfs_output_fy18_1[[1]]
  , dfs_output_fy18_2[[1]]
  , dfs_output_fy18_3[[1]]
  , dfs_output_fy18_4[[1]]
)

save('dfs_output_fy18', file = 'data/output/dfs_output_fy18.Rdata')
```

```{r}
remove(dfs_output_fy18_1, dfs_output_fy18_2, dfs_output_fy18_3, dfs_output_fy18_4)
gc()
```

# Data cleaning

Fill in NAs.

```{r}
dfs_output_fy18 <- dfs_output_fy18 %>%
  # Replace numeric NA with 0
  mutate_if(
    is.numeric
    , ~ replace(., is.na(.), 0)
  )

save('dfs_output_fy18', file = 'data/output/dfs_output_fy18_noNumNA.Rdata')
```

```{r}
dfs_output_fy18 <- dfs_output_fy18 %>%
  # Replace factor NA with missing
  mutate_if(
    is.factor
    , fct_explicit_na
  )

save('dfs_output_fy18', file = 'data/output/dfs_output_fy18_noNA.Rdata')
```

Version without lists.

```{r}
dfs_output_fy18 <- dfs_output_fy18 %>%
  select_if(
    function(x) {!is.list(x)}
  )

save('dfs_output_fy18', file = 'data/output/dfs_output_fy18_noNA_nolists.Rdata')
```

# Response variable

Subsequent year gift amounts (FY19 in this case).

```{r}
fy <- 2019

output_fy18_rv <- catracks$giv_transactions %>%
  group_by(HOUSEHOLD_ID) %>%
  summarise(
    fy19_cash = sum(ifelse(TX_GYPM_IND != 'P' & FISCAL_YEAR == fy, HH_RECOGNITION_CREDIT, 0))
    , fy19_frp = sum(ifelse(TX_GYPM_IND != 'Y' & FISCAL_YEAR == fy, HH_RECOGNITION_CREDIT, 0))
  ) %>%
  group_by()

save('output_fy18_rv', file = 'data/output/output_fy18_rv.Rdata')
```

Check that the HHIDs are unique.

```{r}
output_fy18_rv %>%
  sapply(FUN = function(x) {unique(x) %>% length()})

output_fy18_rv %>% nrow()

dfs_output_fy18$HOUSEHOLD_ID %>% unique() %>% length()

dfs_output_fy18 %>% nrow()
```
